titleSelectors [nfc]: Delete; move only selector to subscriptionSelectors.
As Greg points out [1], this "never made as much sense as the
selectors files like `subscriptionSelectors.js` that are clearly
about a particular part of the app's data model."

[1] https://github.com/zulip/zulip-mobile/pull/4443#discussion_r568190562
