+++ src/selectors.js
@@ -9,7 +9,6 @@ export * from './emoji/emojiSelectors';
 export * from './message/messageSelectors';
 export * from './nav/navSelectors';
 export * from './subscriptions/subscriptionSelectors';
-export * from './title/titleSelectors';
 export * from './topics/topicSelectors';
 export * from './typing/typingSelectors';
 export * from './unread/unreadSelectors';
